+++ Apps/Graphics/TrayApp/CMakeLists.txt
@@ -3,7 +3,7 @@ include(${CMAKE_CURRENT_SOURCE_DIR}/MakeASound.cmake)
 eacp_add_webview_app(TrayApp
         SOURCES     Main.cpp
         API         Api::TrayLauncherApi
-        API_HEADER  Types.h
+        API_HEADER  Api.h
         WEB_DIR     ${CMAKE_CURRENT_SOURCE_DIR}/web
         BUNDLE_ID   com.eacp.trayapp
         BUNDLE_NAME "Tray App"
