+++ Apps/Graphics/TrayApp/Main.cpp
@@ -1,4 +1,4 @@
-#include "Types.h"
+#include "Api.h"
 #include <eacp/Core/Threads/Async.h>
 #include <eacp/Core/App/Clipboard.h>
 #include <eacp/Graphics/Graphics.h>
